requirements.txt
# Data Science & Analysis
pandas>=2.3.0
numpy>=2.2.0
matplotlib>=3.10.0
seaborn>=0.13.0
plotly>=6.4.0
scipy>=1.16.0
scikit-image>=0.25.0

# Image Processing
pillow>=12.0.0
opencv-python>=4.12.0

# Jupyter & Interactive
jupyterlab>=4.4.0
ipython>=9.7.0
rich>=13.7.0

# Web & HTTP
httpx>=0.28.0
beautifulsoup4>=4.14.0

# Browser Automation
selenium>=4.38.0
playwright>=1.56.0

# Document Processing
pytesseract>=0.3.13
pdf2image>=1.17.0
PyPDF2>=3.0.0
python-docx>=1.2.0
openpyxl>=3.1.5
xlsxwriter>=3.2.0
markdown>=3.10
python-pptx>=1.0.0

# Utilities
colorthief>=0.2.1
colorama>=0.4.6
